Polyfills added for homey lower then 2018
