+++ lib/Api.js
@@ -1,5 +1,4 @@
 'use strict';
-require('./polyfills-homey-2018-and-lower'); // Polyfills for Homey 2018 and lower (Node.js 8.x)
 
 const Homey = require('homey');
 const https = require('https');
